bot/requirements.txt
aiohttp==3.8.5
colorama==0.4.6
discord.py==2.3.2
